refactor: 버스 ORM -> 객체 변경 함수 이동
